# crates/sbol-db-server/src/v2/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SynBioHub v2 API",
    "version": "0.1.2",
    "description": "The SynBioHub v2 API: SBOL DB's idiomatic, RESTful presentation of the SynBioHub product surface, mounted under `/api/v2`.\n\nSynBioHub v2 calls the same application services as the SynBioHub v1 compatibility adapter: both expose the same product domain and read and write the same dataset, while retaining separate wire contracts. V2 uses proper HTTP verbs (POST creates, PATCH edits, DELETE deletes, and no GET mutates), JSON request bodies, real pagination with a total, `Accept`-driven content negotiation, a single consistent error envelope, and bearer or same-origin browser-session authentication. V1 preserves classic SynBioHub behavior where compatibility requires it.\n\nWhen public browsing is enabled, a missing or unrecognized resource credential is anonymous and scoped to the public graph. An instance with `requireLogin` rejects anonymous resource requests with `401` while retaining public bootstrap and session routes. A mutating verb attempted anonymously is `403`; invalid login credentials are `401`. Reads are ACL-scoped to the caller's authorized graphs, and an out-of-scope object is a non-disclosing `404`.",
    "contact": {
      "name": "SBOL DB",
      "url": "https://github.com/marpaia/sbol-db"
    },
    "license": {
      "name": "MIT OR Apache-2.0"
    }
  },
  "servers": [
    {
      "url": "/api/v2",
...
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "An `Authorization: Bearer <token>` credential. A missing or unrecognized token is anonymous when public browsing is enabled; `requireLogin` instances reject anonymous resource requests."
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sbol-db-token",
        "description": "The HttpOnly browser session established by `POST /session`. Bearer credentials take precedence. Unsafe cookie-authenticated requests must be same-origin."
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "description": "The single V2 error envelope. Every non-2xx response carries this shape.",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "required": ["code", "message", "status"],
            "properties": {
              "code": { "type": "string", "description": "A stable machine-readable code, shared with the SBOL DB API." },
              "message": { "type": "string", "description": "A human-readable explanation." },
              "status": { "type": "integer", "description": "The HTTP status code, echoed for clients that only read the body." }
            }
          }
        }
      },
      "SearchHit": {
        "type": "object",
        "required": ["uri", "display_id", "version", "name", "description", "object_type", "roles", "owners", "created_at", "modified_at", "score"],
        "properties": {
          "uri": { "type": "string" },
          "display_id": { "type": ["string", "null"] },
          "version": { "type": ["string", "null"] },
          "name": { "type": ["string", "null"] },
          "description": { "type": ["string", "null"] },
          "object_type": { "type": ["string", "null"] },
          "roles": { "type": "array", "items": { "type": "string" } },
          "owners": { "type": "array", "items": { "type": "string" } },
          "created_at": { "type": ["string", "null"], "description": "The earliest visible dcterms:created value." },
          "modified_at": { "type": ["string", "null"], "description": "The latest visible dcterms:modified value." },
          "score": { "type": "number", "description": "The relevance score. Zero for unranked browse-only objects." }
        }
      },
      "SearchResponse": {
        "type": "object",
        "required": ["items", "total", "offset", "limit", "sort", "direction"],
        "properties": {
          "items": { "type": "array", "items": { "$ref": "#/components/schemas/SearchHit" } },
          "total": { "type": "integer", "description": "The total number of in-scope matches, independent of the returned window." },
          "offset": { "type": "integer" },
          "limit": { "type": "integer" },
          "sort": { "type": "string", "enum": ["relevance", "name", "created", "modified", "iri"] },
          "direction": { "type": "string", "enum": ["asc", "desc"] }
        }
      },
      "DiscoveryFacetValue": {
        "type": "object",
        "required": ["iri", "label", "curie", "count"],
        "properties": {
          "iri": { "type": "string" },
          "label": { "type": "string", "description": "Ontology term name when loaded, otherwise a compact IRI fallback." },
          "curie": { "type": ["string", "null"] },
          "count": { "type": "integer", "description": "Distinct visible top-level objects carrying this value." }
        }
      },
      "DiscoveryFacets": {
        "type": "object",
        "required": ["types", "roles"],
        "properties": {
          "types": { "type": "array", "items": { "$ref": "#/components/schemas/DiscoveryFacetValue" } },
          "roles": { "type": "array", "items": { "$ref": "#/components/schemas/DiscoveryFacetValue" } }
        }
      },
      "StructuredSearchRequest": {
        "type": "object",
        "required": ["query"],
        "properties": {
          "strategy": { "type": ["string", "null"], "description": "A registered strategy id. Absent selects the configured default." },
          "query": {
            "type": "object",
            "required": ["kind"],
            "properties": {
              "kind": { "type": "string", "enum": ["text", "similar", "sequence"] },
              "text": { "type": "string" },
              "uri": { "type": "string" },
              "sequence": { "type": "string" },
              "exact": { "type": "boolean", "default": false }
            }
          },
          "filters": {
            "type": "object",
            "properties": {
              "graphs": { "type": "array", "items": { "type": "string" } },
              "object_types": { "type": "array", "items": { "type": "string" } },
              "predicates": { "type": "array", "items": { "type": "object" } }
            }
          },
          "page": {
            "type": "object",
            "properties": {
              "limit": { "type": "integer", "default": 50, "minimum": 1, "maximum": 1000 },
              "cursor": { "type": ["string", "null"] }
            }
          },
          "options": {
            "type": "object",
            "properties": {
              "explain": { "type": "boolean", "default": false },
              "timeout_ms": { "type": ["integer", "null"] }
            }
          }
        }
      },
      "StructuredSearchHit": {
        "type": "object",
        "required": ["document_id", "uri", "score", "score_kind", "object_types"],
        "properties": {
          "document_id": { "type": "string" },
          "uri": { "type": "string" },
          "graph": { "type": ["string", "null"] },
          "score": { "type": "number" },
          "score_kind": {},
          "display_id": { "type": ["string", "null"] },
          "version": { "type": ["string", "null"] },
          "name": { "type": ["string", "null"] },
          "description": { "type": ["string", "null"] },
          "object_types": { "type": "array", "items": { "type": "string" } },
          "evidence": { "type": "array", "items": { "type": "object" } }
        }
      },
      "StructuredSearchPage": {
        "type": "object",
        "required": ["strategy", "items", "total", "execution"],
        "properties": {
          "strategy": {
            "type": "object",
            "required": ["id", "version"],
            "properties": {
              "id": { "type": "string" },
              "version": { "type": "string" }
            }
          },
          "items": { "type": "array", "items": { "$ref": "#/components/schemas/StructuredSearchHit" } },
          "total": {
            "type": "object",
            "required": ["kind"],
            "properties": {
              "kind": { "type": "string", "enum": ["exact", "lower_bound", "unknown"] },
              "value": { "type": "integer" }
            }
          },
          "next_cursor": { "type": ["string", "null"] },
          "execution": { "type": "object" }
        }
      },
...
    },
    "responses": {
      "Error": {
        "description": "An error, in the V2 envelope.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      }
    },
...
  },
  "security": [
    { "bearerAuth": [] },
    { "cookieAuth": [] }
  ],
  "paths": {
...
    "/objects": {
      "get": {
        "tags": ["Objects"],
        "summary": "List objects",
        "description": "The paginated, ACL-scoped object list. Shares the normalized discovery contract and response envelope with `/search`.",
        "operationId": "v2ListObjects",
        "parameters": [
          { "name": "q", "in": "query", "required": false, "description": "A free-text term. Absent browses the whole in-scope corpus.", "schema": { "type": "string" } },
          { "name": "type", "in": "query", "required": false, "description": "Restrict to one rdf:type, given as a full IRI.", "schema": { "type": "string" } },
          { "name": "role", "in": "query", "required": false, "description": "Restrict to one SBOL role, given as a full IRI.", "schema": { "type": "string" } },
          { "name": "collection", "in": "query", "required": false, "description": "Restrict to direct members of this collection IRI.", "schema": { "type": "string" } },
          { "name": "owner", "in": "query", "required": false, "description": "Restrict to objects owned by this user graph IRI.", "schema": { "type": "string" } },
          { "name": "provenance", "in": "query", "required": false, "description": "Case-insensitive substring over mutable provenance.", "schema": { "type": "string" } },
          { "name": "created_after", "in": "query", "required": false, "schema": { "type": "string", "format": "date" } },
          { "name": "created_before", "in": "query", "required": false, "schema": { "type": "string", "format": "date" } },
          { "name": "modified_after", "in": "query", "required": false, "schema": { "type": "string", "format": "date" } },
          { "name": "modified_before", "in": "query", "required": false, "schema": { "type": "string", "format": "date" } },
          { "name": "sort", "in": "query", "required": false, "schema": { "type": "string", "enum": ["relevance", "name", "created", "modified", "iri"], "default": "relevance" } },
          { "name": "direction", "in": "query", "required": false, "description": "Defaults to ascending for name/IRI and descending otherwise.", "schema": { "type": "string", "enum": ["asc", "desc"] } },
          { "name": "offset", "in": "query", "required": false, "schema": { "type": "integer", "minimum": 0, "default": 0 } },
          { "name": "limit", "in": "query", "required": false, "schema": { "type": "integer", "minimum": 1, "maximum": 1000, "default": 50 } }
        ],
        "responses": {
          "200": {
            "description": "A page of ranked objects with a total.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SearchResponse" }
              }
            }
          }
        }
      }
    },
...
    "/search": {
      "get": {
        "tags": ["Search"],
        "summary": "Discover registry objects",
        "description": "Normalized ACL-scoped discovery with text, biological facets, collection/owner/provenance/date narrowing, deterministic sorting, exact totals, and stable paging. The wire uses query parameters rather than V1's path grammar.",
        "operationId": "v2Search",
        "parameters": [
          { "name": "q", "in": "query", "required": false, "description": "The free-text term. Absent browses the whole in-scope corpus.", "schema": { "type": "string" } },
          { "name": "type", "in": "query", "required": false, "description": "Restrict to one rdf:type, given as a full IRI.", "schema": { "type": "string" } },
          { "name": "role", "in": "query", "required": false, "description": "Restrict to one SBOL role, given as a full IRI.", "schema": { "type": "string" } },
          { "name": "collection", "in": "query", "required": false, "description": "Restrict to direct members of this collection IRI.", "schema": { "type": "string" } },
          { "name": "owner", "in": "query", "required": false, "description": "Restrict to objects owned by this user graph IRI.", "schema": { "type": "string" } },
          { "name": "provenance", "in": "query", "required": false, "description": "Case-insensitive substring over mutable provenance.", "schema": { "type": "string" } },
          { "name": "created_after", "in": "query", "required": false, "schema": { "type": "string", "format": "date" } },
          { "name": "created_before", "in": "query", "required": false, "schema": { "type": "string", "format": "date" } },
          { "name": "modified_after", "in": "query", "required": false, "schema": { "type": "string", "format": "date" } },
          { "name": "modified_before", "in": "query", "required": false, "schema": { "type": "string", "format": "date" } },
          { "name": "sort", "in": "query", "required": false, "schema": { "type": "string", "enum": ["relevance", "name", "created", "modified", "iri"], "default": "relevance" } },
          { "name": "direction", "in": "query", "required": false, "description": "Defaults to ascending for name/IRI and descending otherwise.", "schema": { "type": "string", "enum": ["asc", "desc"] } },
          { "name": "offset", "in": "query", "required": false, "schema": { "type": "integer", "minimum": 0, "default": 0 } },
          { "name": "limit", "in": "query", "required": false, "schema": { "type": "integer", "minimum": 1, "maximum": 1000, "default": 50 } }
        ],
        "responses": {
          "200": {
            "description": "A deterministic page of discovery hits with an exact total.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/SearchResponse" } }
            }
          },
          "400": { "$ref": "#/components/responses/Error" },
          "401": { "$ref": "#/components/responses/Error" }
        }
      },
      "post": {
        "tags": ["Search"],
        "summary": "Execute a pluggable search strategy",
        "description": "Executes a structured request against a registered strategy under the caller's ACL graph scope. This is additive to the compatibility GET operation.",
        "operationId": "v2StructuredSearch",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/StructuredSearchRequest" } }
          }
        },
        "responses": {
          "200": {
            "description": "A strategy-labelled page whose total explicitly declares its guarantee.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/StructuredSearchPage" } }
            }
          },
          "400": { "$ref": "#/components/responses/Error" },
          "503": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
    "/search/facets": {
      "get": {
        "tags": ["Search"],
        "summary": "Discover visible biological facets",
        "description": "Returns exact distinct-subject counts for rdf:type and SBOL role values under the caller's ACL scope, enriched with ontology labels when available.",
        "operationId": "v2SearchFacets",
        "responses": {
          "200": {
            "description": "Visible type and role values ordered by count and label.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/DiscoveryFacets" } }
            }
          },
          "401": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
  }
}